gui: Protect against impossible NRE
This really should not be possible - in the callback for getting
preferred width the natural_size parameter is received as a null
reference, causing a runtime exception attempting to set it.

We add some unsafe blocks and check the pointer value before
dereferencing each of the out parameters we receive in both the size
calculation callback functions although we only saw it with width and
the natural_size paramter.